requirements.txt
altgraph==0.17.3
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.8
Django==4.2.6
djangorestframework==3.14.0
exceptiongroup==1.1.1
filelock==3.13.1
h11==0.14.0
idna==3.4
numpy==1.25.1
outcome==1.2.0
packaging==23.1
pefile==2023.2.7
Pillow==10.0.0
platformdirs==4.2.0
pycparser==2.21
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.25.0
webdriver-manager==3.8.6
wsproto==1.2.0
wxPython==4.2.1
